Add base login page
