+++ TremorTrainer/TremorTrainer/ViewModels/GetStartedViewModel.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+using TremorTrainer.Views;
+using Xamarin.Forms;
+
+namespace TremorTrainer.ViewModels
+{
+    public class GetStartedViewModel : BaseViewModel
+    {
+        public Command PrescribedCommand { get; }
+        public Command MaintenanceCommand { get; }
+
+
+        public GetStartedViewModel()
+        {
+            PrescribedCommand = new Command(OnDay1SessionClicked);
+            MaintenanceCommand = new Command(OnSubsequentDaysButtonClicked);
+        }
+
+        private static async void OnDay1SessionClicked(object obj)
+        {
+            App.Current.Properties["IsInductionSession"] = true;
+            await Shell.Current.GoToAsync("//AccelerometerPage");
+
+        }
+
+        private static async void OnSubsequentDaysButtonClicked(object obj)
+        {
+            App.Current.Properties["IsInductionSession"] = false;
+            await Shell.Current.GoToAsync("//AccelerometerPage");
+        }
+    }
+}
