+++ TremorTrainer/TremorTrainer/Views/LoginPage.xaml
@@ -1,0 +1,87 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage
+    xmlns="http://xamarin.com/schemas/2014/forms"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:buttons="clr-namespace:Syncfusion.XForms.Buttons;assembly=Syncfusion.Buttons.XForms"
+    xmlns:inputLayout="clr-namespace:Syncfusion.XForms.TextInputLayout;assembly=Syncfusion.Core.XForms"
+    xmlns:core="clr-namespace:Syncfusion.XForms.Core;assembly=Syncfusion.Core.XForms"
+    xmlns:gradient="clr-namespace:Syncfusion.XForms.Graphics;assembly=Syncfusion.Core.XForms"
+    xmlns:viewmodels="clr-namespace:TremorTrainer.ViewModels"
+    x:DataType="viewmodels:LoginViewModel"
+    x:Class="TremorTrainer.Views.LoginPage"
+    NavigationPage.HasNavigationBar="False">
+
+        <Grid>
+            <gradient:SfGradientView Style="{StaticResource BlueGradientStyle}" />
+
+            <ScrollView HorizontalOptions="Center">
+                <ScrollView.VerticalOptions>
+                    <OnIdiom
+                        x:TypeArguments="LayoutOptions"
+                        Default="CenterAndExpand"
+                        Phone="FillAndExpand" />
+                </ScrollView.VerticalOptions>
+
+                <StackLayout
+                    Margin="{core:OnPlatformOrientationThickness PhonePortrait='20,32',
+                                                                 PhoneLandscape='150,32',
+                                                                 TabletPortrait='200,50',
+                                                                 TabletLandscape='300,50',
+                                                                 Desktop='30'}"
+                    Spacing="0"
+                    VerticalOptions="FillAndExpand"
+                    WidthRequest="{OnPlatform Default='-1',
+                                              UWP='350'}">
+
+                    <Grid
+                        RowDefinitions="Auto, Auto, Auto, Auto, Auto, Auto, Auto, Auto"
+                        RowSpacing="0"
+                        VerticalOptions="CenterAndExpand">
+
+                        <!--  Label to display the title of this page  -->
+                        <Label
+                            FontSize="20"
+                            Style="{StaticResource TitleLabelStyle}"
+                            Text="Welcome to the Tremor Retrainer project"
+                            TextColor="{DynamicResource Gray-White}" />
+
+                        <!--  Label to display the description of this page  -->
+                        <Label
+                            Grid.Row="1"
+                            Margin="0,8"
+                            Style="{StaticResource DescriptionLabelStyle}"
+                            Text="Let's get to work"
+                            TextColor="{DynamicResource Gray-White}" />
+                    <inputLayout:SfTextInputLayout
+                        Grid.Row="4"
+                        Hint="Username"
+                        HintLabelStyle="{DynamicResource Gray-White}">
+
+                        <Entry
+                            TextColor="{DynamicResource Gray-White}"
+                            />
+                    </inputLayout:SfTextInputLayout>
+                    <inputLayout:SfTextInputLayout
+                        Grid.Row="5"
+                        EnablePasswordVisibilityToggle="true"
+                        Hint="Password">
+                        <Entry
+                            IsPassword="True"/>
+                    </inputLayout:SfTextInputLayout>
+
+                        <!--  Prescribed Session Button  -->
+                        <buttons:SfButton
+                            Grid.Row="6"
+                            Margin="0,16"
+                            Command="{Binding LoginCommand}"
+                            HorizontalOptions="Fill"
+                            Style="{StaticResource GradientButtonStyle}"
+                            Text="Login" 
+                        />
+
+                    </Grid>
+                </StackLayout>
+            </ScrollView>
+
+        </Grid>
+    </ContentPage>
